Add: Spare immer das Zehnfache des Stadtwache Stundenlohns.

+++ src/main/java/de/binary101/core/data/character/Attribute.java
@@ -8,5 +8,5 @@ public class Attribute {
 	@Getter @Setter private AttributeEnum type;
 	@Getter @Setter private int baseValue;
 	@Getter @Setter private int bonusValue;
-	@Getter @Setter private double priceForNextUpgrade;
+	@Getter @Setter private long priceForNextUpgrade;
 }
